match any complete path segment
fixes cases where a correct path might be given, but it doesn't match up with the `file.path` or `file.relative`. For example, the path might be `a/b/c/d/e/f.hbs`, and relative might be `e/f.hbs`, and the user provides `c/d/e/f.hbs`, which is correct, but would not have matched before.